+++ fastai/imports/core.py
@@ -2,7 +2,7 @@ import csv, gc, gzip, os, pickle, shutil, sys, warnings, yaml
 import math, matplotlib.pyplot as plt, numpy as np, pandas as pd, random
 import scipy.stats, scipy.special
 import abc, collections, hashlib, itertools, json, operator, pathlib
-import mimetypes, inspect, typing, functools
+import mimetypes, inspect, typing, functools, importlib
 import html, re, spacy, requests, tarfile, numbers
 
 from abc import abstractmethod, abstractproperty